moved all property metadata into the public module
